Use device mode specific share icon
Resolves #1325
