+++ src/features/comment/useCommentActions.ts
@@ -9,7 +9,6 @@ import {
   flagOutline,
   pencilOutline,
   personOutline,
-  shareOutline,
   textOutline,
   trashOutline,
 } from "ionicons/icons";
@@ -45,6 +44,7 @@ import { isDownvoteEnabledSelector } from "../auth/siteSlice";
 import { compact } from "lodash";
 import { isStubComment } from "./CommentHeader";
 import { getVoteErrorMessage } from "../../helpers/lemmyErrors";
+import { getShareIcon } from "../../helpers/device";
 
 export interface CommentActionsProps {
   comment: CommentView | PersonMentionView | CommentReplyView;
@@ -262,7 +262,7 @@ export default function useCommentActions({
         },
         {
           text: "Share",
-          icon: shareOutline,
+          icon: getShareIcon(),
           handler: () => {
             share(comment);
           },
