+++ src/features/post/actions/PostActions.tsx
@@ -1,11 +1,12 @@
 import { styled } from "@linaria/react";
 import { IonIcon } from "@ionic/react";
-import { arrowUndoOutline, linkOutline, shareOutline } from "ionicons/icons";
+import { arrowUndoOutline, linkOutline } from "ionicons/icons";
 import { PostView } from "lemmy-js-client";
 import { VoteButton } from "../shared/VoteButton";
 import { ActionButton } from "./ActionButton";
 import { SaveButton } from "../shared/SaveButton";
 import { share } from "../../../helpers/lemmy";
+import { getShareIcon } from "../../../helpers/device";
 
 const Container = styled.div`
   display: flex;
@@ -42,7 +43,7 @@ export default function PostActions({ post, onReply }: PostActionsProps) {
         <IonIcon icon={arrowUndoOutline} />
       </ActionButton>
       <ActionButton>
-        <IonIcon icon={shareOutline} onClick={() => share(post.post)} />
+        <IonIcon icon={getShareIcon()} onClick={() => share(post.post)} />
       </ActionButton>
     </Container>
   );
